+++ KeViraKombinaTodos.Core/Models/Produto.cs
@@ -9,8 +9,8 @@ namespace KeViraKombinaTodos.Core.Models
         public int ProdutoID { get; set; }
         public string Descricao { get; set; }
         public string Codigo { get; set; }
-        public double Valor { get; set; }
-        public double Quantidade { get; set; }
+        public decimal Valor { get; set; }
+        public decimal Quantidade { get; set; }
         public bool Ativo { get; set; }
         public DateTime? DataCriacao { get; set; }
         public DateTime? DataModificacao { get; set; }
